Files in static folder are not copied.
First of all: broccoli-static-compiler is deprecated by broccoli-funnel.
That is why I have replaced it in the dependencies.

However the reason for this patch is that, in my project, the static
folder was empty.

I'm not sure if this is a bug in broccoli-funnel. However it looks like
subsequent calls to the filter, inherit the options used in previous
calls. We need then to be explicit on what to include because a filter
on file names was set by previous calls to the filter.